+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Error on fetching experimental releases due to change in the structure of their web pages.
+- Lag on settings tab due to calculating disk space on the GUI thread.
 - Typos.
 
 ## [0.4.1] - 2021-11-18: Never rush a release